refactor: sign-in and sign-up forms
