+++ src/shared/constants/index.ts
@@ -6,3 +6,9 @@ export {
   XL2_BREAKPOINT,
   XS_BREAKPOINT,
 } from './breakpoints'
+
+export { EMPTY_STRING } from './base'
+
+export { USERNAME_PATTERN, PASSWORD_PATTERN } from './regexs'
+
+export { AppRoutes, AuthRoutes, GeneralRoutes } from './routes'
