+++ src/shared/lib/hooks/index.ts
@@ -1,2 +1,4 @@
 export { useResponsive } from './use-responsive'
-export { useLoader } from '@/shared/lib/hooks/use-loader'
+export { useLoader } from './use-loader'
+export { useTranslation } from './use-translation'
+export { useFormRevalidateWithLocale } from './use-form-revalidate-with-locale'
