+++ src/shared/types/index.ts
@@ -3,3 +3,4 @@ export type { ReturnComponent } from '@/shared/types/return-component'
 export type { Nullable } from 'src/shared/types/nullable'
 export type { PolymorphComponentProps } from 'src/shared/types/polymorph'
 export type { PolymorphComponentPropsWithRef } from 'src/shared/types/polymorph'
+export type { UseFormRef } from './form'
